#1
In the modified program, ensemble methods were used to improve the accuracy of the predictions. Instead of using the best individual in the population to make a prediction, the program now generates multiple sets of weights using the genetic algorithm, and then combines the predictions of each set of weights to generate a final prediction. This helps to reduce the impact of random fluctuations in the fitness landscape and improves the robustness of the predictions.